refactor: extract `middleware.py`
